requirements.txt
Flask==2.0.1
pymongo==3.12.0
Flask-PyMongo==2.3.0
suds-py3==1.4.4.1